combined NewComment and ListComment into one servlet /comments
